added big integer for table

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,13 +15,13 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->engine = 'InnoDB';
-            $table->integer('id', true);
+            $table->bigInteger('id', true);
             $table->string('email')->unique();
             $table->date('birthday');
             $table->string('first_name', 30);
             $table->string('last_name', 30);
             $table->string('gender', 10);
-            $table->integer('facebook_id');
+            $table->bigInteger('facebook_id');
             $table->string('profile_picture', 255);
             $table->timestamps();
         });
